image-encryption (4)/image-encryption/requirements.txt
ttkthemes~=3.2.2
pycryptodome